refactor: Exception 리팩토링
- 에러 코드 교체
- 사용하지 않는 Exception 제거
