+++ swith-me/swith-me-notification-api/src/main/java/project/swithme/notification/core/web/exception/InvalidOrderException.java
@@ -1,11 +1,11 @@
 package project.swithme.notification.core.web.exception;
 
-import static project.study.support.codeandmessage.order.OrderCodeAndMessage.INVALID_ORDER;
+import static project.study.support.codeandmessage.order.OrderErrorCodeAndMessage.ORDER_NOT_FOUND;
 import project.study.support.exception.DomainException;
 
 public class InvalidOrderException extends DomainException {
 
     public InvalidOrderException() {
-        super(INVALID_ORDER);
+        super(ORDER_NOT_FOUND);
     }
 }
