+++ swith-me/swith-me-notification-api/src/main/java/project/swithme/notification/core/web/exception/NotificationNotFoundException.java
@@ -1,6 +1,6 @@
 package project.swithme.notification.core.web.exception;
 
-import static project.study.support.codeandmessage.notification.NotificationCodeAndMessage.NOTIFICATION_NOT_FOUND;
+import static project.study.support.codeandmessage.notification.NotificationErrorCodeAndMessage.NOTIFICATION_NOT_FOUND;
 import project.study.support.exception.DomainException;
 
 public class NotificationNotFoundException extends DomainException {
