+++ swith-me/swith-me-order-api/src/main/java/project/swithme/order/core/exception/OrderNotFoundException.java
@@ -1,6 +1,6 @@
 package project.swithme.order.core.exception;
 
-import static project.study.support.codeandmessage.order.OrderCodeAndMessage.ORDER_NOT_FOUND;
+import static project.study.support.codeandmessage.order.OrderErrorCodeAndMessage.ORDER_NOT_FOUND;
 import project.study.support.exception.DomainException;
 
 public class OrderNotFoundException extends DomainException {
