--- swith-me/swith-me-payment-api/src/main/java/project/swithme/payment/core/exception/InvalidPriceException.java
@@ -1,11 +1,0 @@
-package project.swithme.payment.core.exception;
-
-import static project.study.support.codeandmessage.order.OrderCodeAndMessage.INVALID_PRICE;
-import project.study.support.exception.DomainException;
-
-public class InvalidPriceException extends DomainException {
-
-    public InvalidPriceException() {
-        super(INVALID_PRICE);
-    }
-}
